Added policies for amenities

+++ app/Http/Controllers/Amenity/DestroyController.php
@@ -8,8 +8,10 @@ class DestroyController extends BaseController
 {
     public function __invoke(Amenity $amenity)
     {
+        $this->authorize('delete', $amenity);
+
         $this->service->destroy($amenity);
-        
+
         return redirect()->route('amenity.index');
     }
 }
